Add support for adding dynamic storage providers for a Deployment
