+++ frontend/packages/ceph-storage-plugin/locales/en/ceph-storage-plugin.json
@@ -1,6 +1,11 @@
 {
   "Troubleshoot": "Troubleshoot",
   "Add Capacity": "Add Capacity",
+  "Use existing claim": "Use existing claim",
+  "Select claim": "Select claim",
+  "Create new claim": "Create new claim",
+  "Create": "Create",
+  "Cancel": "Cancel",
   "{{replica}} Replication": "{{replica}} Replication",
   "Pool name": "Pool name",
   "my-block-pool": "my-block-pool",
@@ -18,8 +23,6 @@
   "Pool creation is not supported for openshift container storage external mode.": "Pool creation is not supported for openshift container storage external mode.",
   "A block pool is a logical entity providing elastic capacity to applications and workloads. Pools provide a means of supporting policies for access data resilience and storage efficiency.": "A block pool is a logical entity providing elastic capacity to applications and workloads. Pools provide a means of supporting policies for access data resilience and storage efficiency.",
   "Block Pool Creation Form": "Block Pool Creation Form",
-  "Create": "Create",
-  "Cancel": "Cancel",
   "Name": "Name",
   "Bucket Name": "Bucket Name",
   "Type": "Type",
@@ -343,16 +346,14 @@
   "Switch to Credentials": "Switch to Credentials",
   "Access Key Field": "Access Key Field",
   "Secret Key Field": "Secret Key Field",
-  "Type: Namespace | Policy: {{policyType}}": "Type: Namespace | Policy: {{policyType}}",
-  "Type: Standard | Tiers: {{tiers}}": "Type: Standard | Tiers: {{tiers}}",
-  "Create Object Bucket Claim": "Create Object Bucket Claim",
-  "Edit YAML": "Edit YAML",
-  " Object Bucket Claim Name": " Object Bucket Claim Name",
+  " ObjectBucketClaim Name": " ObjectBucketClaim Name",
   "my-object-bucket": "my-object-bucket",
   "If not provided a generic name will be generated.": "If not provided a generic name will be generated.",
   "Defines the object-store service and the bucket provisioner.": "Defines the object-store service and the bucket provisioner.",
   "BucketClass": "BucketClass",
   "Select BucketClass": "Select BucketClass",
+  "Create ObjectBucketClaim": "Create ObjectBucketClaim",
+  "Edit YAML": "Edit YAML",
   "Attach to Deployment": "Attach to Deployment",
   "Object Bucket Claim Details": "Object Bucket Claim Details",
   "Object Bucket": "Object Bucket",
