+++ frontend/packages/ceph-storage-plugin/src/components/attach-obc/attach-obc-deployment.scss
@@ -1,0 +1,4 @@
+.ceph-attach-obc__subgroup {
+  margin-left: var(--pf-global--spacer--md);
+  margin-top: var(--pf-global--spacer--md);
+}
